Refactoring of getContactParentHierarchy
